create party_result_by_department_business class and its test
