+++ app/adapters/dependency/dependency_business.py
@@ -2,7 +2,7 @@ from app.domain.business.candidate.candidate_business import CandidateBusiness
 from app.domain.business.department_business import DepartmentBusiness
 from app.domain.business.district_business import DistrictBusiness
 from app.domain.business.deputy_business import DeputyBusiness
-from app.domain.business.party_business import PartyBusiness
+from app.domain.business.party.party_business import PartyBusiness
 from app.domain.business.result_business import ResultBusiness
 from app.adapters.driven.MySql.cache import Cache
 from app.adapters.driven.MySql.my_sql_candidate_repository import MySqlCandidateRepository
@@ -21,7 +21,7 @@ class DependencyBusiness :
         deputy_repo = MySqlDeputyRepository(cache)
         result_repo = MySqlResultRepository(cache)
         district_repo = MySqlDistrictRepository(cache)
-        self.party_business = PartyBusiness(party_repo, candidate_repo, district_repo, department_repo)
+        self.party_business = PartyBusiness(party_repo, candidate_repo)
         self.result_business = ResultBusiness(result_repo)
         self.department_business = DepartmentBusiness(department_repo, district_repo, result_repo)
         self.district_business = DistrictBusiness(district_repo, self.department_business, self.result_business)
