+++ app/domain/business/party/party_result_by_department_business.py
@@ -1,7 +1,6 @@
-from app.domain.business.candidate.sortered_candidate import SorteredCandidate
 from app.domain.factory.factorydepartmentpartyresult import FactoryDepartmentPartyResult
 
-class PartyBusiness :
+class PartyResultByDepartmentBusiness : 
     def __init__(self, party_repo, candidate_repo, district_repo, department_repo) -> None:
         self.party_repo = party_repo
         self.candidate_repo = candidate_repo
@@ -9,44 +8,6 @@ class PartyBusiness :
         self.department_repo = department_repo
         self.factory_department_party_result = FactoryDepartmentPartyResult()
         
-    
-    def get_parties(self) : 
-        parties = self.party_repo.get_parties()
-        return parties
-    
-    
-    def get_party_by_short_name(self, short_name):
-        party_by_short_name = None
-        parties = self.party_repo.get_parties()
-        for party in parties : 
-            if party.short_name == short_name : 
-                party_by_short_name = party
-                break
-        return party_by_short_name
-    
-    
-    def get_top_candidates_for_each_party_all_rounds(self, limit) :
-        all_candidates = self.candidate_repo.get_candidates()
-        all_parties = self.party_repo.get_parties()
-        all_candidates_by_party = {}
-        
-        for candidate in all_candidates :
-            for party in all_parties : 
-                if candidate.party_id == party.id : 
-                    if (party.short_name in all_candidates_by_party) == False:
-                        all_candidates_by_party[party.short_name] = []
-                    all_candidates_by_party[party.short_name].append(candidate)
-                    break
-
-        for party in all_parties :
-            candidates = all_candidates_by_party[party.short_name]
-            sortered  = SorteredCandidate(candidates, all_parties)
-            candidates_sortered = sortered.sort_all_candidates(limit)
-            all_candidates_by_party[party.short_name] = candidates_sortered        
-            
-        return all_candidates_by_party
-    
-    #externalize in a specific class
     def get_top_departments_for_each_party_all_rounds(self) :
         all_candidates = self.candidate_repo.get_candidates()
         all_parties = self.party_repo.get_parties()
