+++ tests/domain/factory/test_factory_department_party_result.py
@@ -1,18 +1,18 @@
 import unittest
 
-from app.domain.factory.factorydepartmentpartyresult import FactoryDeartmentPartyResult
+from app.domain.factory.factorydepartmentpartyresult import FactoryDepartmentPartyResult
 from tests.assert_test import AssertTest
 from tests.faker.faker_candidate import getCandidates_by_partys_departments
 from tests.faker.faker_department import getDepartments_by_ids
 
-class FactoryDepartmentPartyResult(unittest.TestCase) :
+class FactoryDepartmentPartyResultTest(unittest.TestCase) :
     def __init__(self, methodName: str = "runTest") -> None:
         super().__init__(methodName)
         self.assert_test = AssertTest(self)
         
     
     def test_factory_department_party_result(self) : 
-        factory_department_party_result = FactoryDeartmentPartyResult()
+        factory_department_party_result = FactoryDepartmentPartyResult()
         
         dep_party_result = factory_department_party_result.construct_department_party_result(650, "Gironde", 33, 66.67)
         
@@ -21,7 +21,7 @@ class FactoryDepartmentPartyResult(unittest.TestCase) :
         
         
     def test_factory_department_party_result_from_candidates_list(self) : 
-        factory_department_party_result = FactoryDeartmentPartyResult()
+        factory_department_party_result = FactoryDepartmentPartyResult()
         candidates = getCandidates_by_partys_departments(["ENS"], [33])
         first_round_voting = []
         for candidate in candidates :
